Use new version of afterburner
