+++ build.gradle
@@ -123,7 +123,7 @@ dependencies {
     compile 'com.google.guava:guava:24.1-jre'
 
     // JavaFX stuff
-    compile 'com.airhacks:afterburner.fx:1.7.0'
+    //compile 'com.airhacks:afterburner.fx:1.7.0'
     compile 'de.codecentric.centerdevice:javafxsvg:1.3.0'
     compile 'de.jensd:fontawesomefx-materialdesignfont:1.7.22-4'
     compile 'de.saxsys:mvvmfx-validation:1.7.0'
@@ -132,6 +132,7 @@ dependencies {
     compile 'org.fxmisc.flowless:flowless:0.6'
     compile 'org.fxmisc.richtext:richtextfx:0.8.2'
     compile 'com.sibvisions.external.jvxfx:dndtabpane:0.1'
+    compile 'javax.inject:javax.inject:1'
 
 
     // Cannot be updated to 9.*.* until Jabref works with Java 9
