+++ src/main/java/org/jabref/gui/documentviewer/DocumentViewer.fxml
@@ -15,7 +15,7 @@
 <?import de.jensd.fx.glyphs.materialdesignicons.MaterialDesignIconView?>
 <DialogPane xmlns:fx="http://javafx.com/fxml/1" depthTest="DISABLE" prefHeight="600.0" prefWidth="750.0"
             xmlns="http://javafx.com/javafx/8.0.112"
-            fx:controller="org.jabref.gui.documentviewer.DocumentViewerController">
+            fx:controller="org.jabref.gui.documentviewer.DocumentViewerView">
     <content>
         <BorderPane fx:id="mainPane">
             <fx:define>
