+++ src/main/java/org/jabref/gui/documentviewer/DocumentViewerView.java
@@ -1,25 +1,129 @@
 package org.jabref.gui.documentviewer;
 
-import javafx.scene.control.Alert.AlertType;
-import javafx.scene.control.ButtonBar;
-import javafx.scene.control.DialogPane;
+import javax.inject.Inject;
 
-import org.jabref.gui.AbstractDialogView;
-import org.jabref.gui.FXDialog;
+import javafx.event.ActionEvent;
+import javafx.fxml.FXML;
+import javafx.scene.Parent;
+import javafx.scene.control.ComboBox;
+import javafx.scene.control.Dialog;
+import javafx.scene.control.Label;
+import javafx.scene.control.ScrollBar;
+import javafx.scene.control.TextField;
+import javafx.scene.control.ToggleButton;
+import javafx.scene.layout.BorderPane;
+import javafx.stage.Modality;
+
+import org.jabref.gui.StateManager;
+import org.jabref.gui.util.OnlyIntegerFormatter;
+import org.jabref.gui.util.TaskExecutor;
+import org.jabref.gui.util.ViewModelListCellFactory;
 import org.jabref.logic.l10n.Localization;
+import org.jabref.model.entry.LinkedFile;
 
-public class DocumentViewerView extends AbstractDialogView {
+import com.airhacks.afterburner.views.ViewLoader;
 
-    @Override
-    public void show() {
-        FXDialog dialog = new FXDialog(AlertType.INFORMATION, Localization.lang("Document viewer"), false);
-        DialogPane dialogPane = (DialogPane) this.getView();
+public class DocumentViewerView extends Dialog<Void> {
 
-        // Remove button bar at bottom
-        dialogPane.getChildren().removeIf(node -> node instanceof ButtonBar);
+    @FXML private ScrollBar scrollBar;
+    @FXML private ComboBox<LinkedFile> fileChoice;
+    @FXML private BorderPane mainPane;
+    @FXML private ToggleButton modeLive;
+    @FXML private TextField currentPage;
+    @FXML private Label maxPages;
 
-        dialog.setDialogPane(dialogPane);
-        dialog.setResizable(true);
-        dialog.show();
+    @Inject private StateManager stateManager;
+    @Inject private TaskExecutor taskExecutor;
+    private DocumentViewerControl viewer;
+    private DocumentViewerViewModel viewModel;
+
+    public DocumentViewerView() {
+        this.setTitle(Localization.lang("Document viewer"));
+        this.initModality(Modality.NONE);
+        this.setResizable(true);
+
+        Parent content = ViewLoader.view(this)
+                                   .load()
+                                   .getView();
+        this.getDialogPane().setContent(content);
+
+        //TODO?: Remove button bar at bottom
+        //dialogPane.getChildren().removeIf(node -> node instanceof ButtonBar);
+    }
+
+    @FXML
+    private void initialize() {
+        viewModel = new DocumentViewerViewModel(stateManager);
+
+        setupViewer();
+        setupScrollbar();
+        setupFileChoice();
+        setupPageControls();
+        setupModeButtons();
+    }
+
+    private void setupModeButtons() {
+        viewModel.liveModeProperty().bind(modeLive.selectedProperty());
+    }
+
+    private void setupScrollbar() {
+        scrollBar.valueProperty().bindBidirectional(viewer.scrollYProperty());
+        scrollBar.maxProperty().bind(viewer.scrollYMaxProperty());
+    }
+
+    private void setupPageControls() {
+        OnlyIntegerFormatter integerFormatter = new OnlyIntegerFormatter(1);
+        viewModel.currentPageProperty().bindBidirectional(integerFormatter.valueProperty());
+        currentPage.setTextFormatter(integerFormatter);
+        maxPages.textProperty().bind(viewModel.maxPagesProperty().asString());
+    }
+
+    private void setupFileChoice() {
+        ViewModelListCellFactory<LinkedFile> cellFactory = new ViewModelListCellFactory<LinkedFile>()
+                .withText(LinkedFile::getLink);
+        fileChoice.setButtonCell(cellFactory.call(null));
+        fileChoice.setCellFactory(cellFactory);
+        fileChoice.getSelectionModel().selectedItemProperty().addListener(
+                (observable, oldValue, newValue) -> viewModel.switchToFile(newValue));
+        // We always want that the first item is selected after a change
+        // This also automatically selects the first file on the initial load
+        fileChoice.itemsProperty().addListener(
+                (observable, oldValue, newValue) -> fileChoice.getSelectionModel().selectFirst());
+        fileChoice.itemsProperty().bind(viewModel.filesProperty());
+    }
+
+    private void setupViewer() {
+        viewer = new DocumentViewerControl(taskExecutor);
+        viewModel.currentDocumentProperty().addListener((observable, oldDocument, newDocument) -> {
+            if (newDocument != null) {
+                viewer.show(newDocument);
+            }
+        });
+        viewModel.currentPageProperty().bindBidirectional(viewer.currentPageProperty());
+        mainPane.setCenter(viewer);
+    }
+
+    public void nextPage(ActionEvent actionEvent) {
+        viewModel.showNextPage();
+    }
+
+    public void previousPage(ActionEvent actionEvent) {
+        viewModel.showPreviousPage();
+    }
+
+    public void fitWidth(ActionEvent actionEvent) {
+        viewer.setPageWidth(viewer.getWidth());
+    }
+
+    public void zoomIn(ActionEvent actionEvent) {
+        viewer.changePageWidth(100);
+    }
+
+    public void zoomOut(ActionEvent actionEvent) {
+        viewer.changePageWidth(-100);
+    }
+
+    public void fitSinglePage(ActionEvent actionEvent) {
+        viewer.setPageHeight(viewer.getHeight());
     }
 }
