+++ src/main/java/org/jabref/gui/entryeditor/fileannotationtab/FileAnnotationTab.java
@@ -1,5 +1,6 @@
 package org.jabref.gui.entryeditor.fileannotationtab;
 
+import javafx.scene.Parent;
 import javafx.scene.control.Tooltip;
 
 import org.jabref.gui.entryeditor.EntryEditorTab;
@@ -7,6 +8,8 @@ import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.pdf.FileAnnotationCache;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.FieldName;
+
+import com.airhacks.afterburner.views.ViewLoader;
 
 public class FileAnnotationTab extends EntryEditorTab {
 
@@ -26,6 +29,9 @@ public class FileAnnotationTab extends EntryEditorTab {
 
     @Override
     protected void bindToEntry(BibEntry entry) {
-        setContent(new FileAnnotationTabView(entry, fileAnnotationCache).getView());
+        Parent content = ViewLoader.view(new FileAnnotationTabView(entry, fileAnnotationCache))
+                                   .load()
+                                   .getView();
+        setContent(content);
     }
 }
