+++ src/main/java/org/jabref/gui/errorconsole/ErrorConsole.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.control.ListView?>
 <?import javafx.scene.layout.BorderPane?>
 <DialogPane xmlns:fx="http://javafx.com/fxml/1" prefHeight="500.0" prefWidth="1080.0"
-            xmlns="http://javafx.com/javafx/8.0.60" fx:controller="org.jabref.gui.errorconsole.ErrorConsoleController">
+            xmlns="http://javafx.com/javafx/8.0.60" fx:controller="org.jabref.gui.errorconsole.ErrorConsoleView">
     <content>
         <BorderPane>
             <bottom>
