+++ src/main/java/org/jabref/gui/help/AboutDialogView.java
@@ -1,19 +1,89 @@
 package org.jabref.gui.help;
 
-import javafx.scene.control.Alert.AlertType;
-import javafx.scene.control.DialogPane;
+import javax.inject.Inject;
 
-import org.jabref.gui.AbstractDialogView;
-import org.jabref.gui.FXDialog;
+import javafx.fxml.FXML;
+import javafx.scene.Parent;
+import javafx.scene.control.Dialog;
+import javafx.scene.control.TextArea;
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
+
+import org.jabref.gui.ClipBoardManager;
+import org.jabref.gui.DialogService;
 import org.jabref.logic.l10n.Localization;
+import org.jabref.logic.util.BuildInfo;
 
-public class AboutDialogView extends AbstractDialogView {
+import com.airhacks.afterburner.views.ViewLoader;
+import de.codecentric.centerdevice.javafxsvg.SvgImageLoaderFactory;
 
-    @Override
-    public void show() {
-        FXDialog aboutDialog = new FXDialog(AlertType.INFORMATION, Localization.lang("About JabRef"));
-        aboutDialog.setDialogPane((DialogPane) this.getView());
-        aboutDialog.show();
+public class AboutDialogView extends Dialog<Void> {
+
+    @FXML protected ImageView iconImage;
+    @Inject private DialogService dialogService;
+    @Inject private ClipBoardManager clipBoardManager;
+    @Inject private BuildInfo buildInfo;
+
+    @FXML private TextArea textAreaVersions;
+    private AboutDialogViewModel viewModel;
+
+    public AboutDialogView() {
+        this.setTitle(Localization.lang("About JabRef"));
+
+        Parent content = ViewLoader.view(this)
+                                   .load()
+                                   .getView();
+        this.getDialogPane().setContent(content);
+    }
+
+    @FXML
+    private void initialize() {
+        viewModel = new AboutDialogViewModel(dialogService, clipBoardManager, buildInfo);
+
+        SvgImageLoaderFactory.install();
+        Image icon = new Image(this.getClass().getResourceAsStream("/icons/jabref.svg"));
+        iconImage.setImage(icon);
+        textAreaVersions.setText(viewModel.getVersionInfo());
+    }
+
+    @FXML
+    private void closeAboutDialog() {
+        close();
+    }
+
+    @FXML
+    private void copyVersionToClipboard() {
+        viewModel.copyVersionToClipboard();
+    }
+
+    @FXML
+    private void openJabrefWebsite() {
+        viewModel.openJabrefWebsite();
+    }
+
+    @FXML
+    private void openExternalLibrariesWebsite() {
+        viewModel.openExternalLibrariesWebsite();
+    }
+
+    @FXML
+    private void openGithub() {
+        viewModel.openGithub();
+    }
+
+    @FXML
+    public void openChangeLog() {
+        viewModel.openChangeLog();
+    }
+
+    @FXML
+    public void openLicense() {
+        viewModel.openLicense();
+    }
+
+    @FXML
+    public void openDonation() {
+        viewModel.openDonation();
     }
 
 }
