+++ src/main/java/org/jabref/gui/keyboard/KeyBindingsDialogView.java
@@ -1,23 +1,89 @@
 package org.jabref.gui.keyboard;
 
-import javafx.scene.control.Alert.AlertType;
-import javafx.scene.control.DialogPane;
+import javax.inject.Inject;
+
+import javafx.fxml.FXML;
+import javafx.scene.Parent;
+import javafx.scene.control.Dialog;
+import javafx.scene.control.SelectionMode;
+import javafx.scene.control.SelectionModel;
+import javafx.scene.control.TreeItem;
+import javafx.scene.control.TreeTableColumn;
+import javafx.scene.control.TreeTableView;
 import javafx.stage.Stage;
 
-import org.jabref.gui.AbstractDialogView;
-import org.jabref.gui.FXDialog;
+import org.jabref.gui.DialogService;
+import org.jabref.gui.JabRefIcon;
+import org.jabref.gui.util.RecursiveTreeItem;
+import org.jabref.gui.util.ViewModelTreeTableCellFactory;
 import org.jabref.logic.l10n.Localization;
+import org.jabref.preferences.PreferencesService;
 
-public class KeyBindingsDialogView extends AbstractDialogView {
+import com.airhacks.afterburner.views.ViewLoader;
+import org.fxmisc.easybind.EasyBind;
 
-    @Override
-    public void show() {
-        FXDialog keyBindingsDialog = new FXDialog(AlertType.INFORMATION, Localization.lang("Key bindings"));
-        keyBindingsDialog.setDialogPane((DialogPane) this.getView());
-        keyBindingsDialog.setResizable(true);
-        ((Stage) keyBindingsDialog.getDialogPane().getScene().getWindow()).setMinHeight(475);
-        ((Stage) keyBindingsDialog.getDialogPane().getScene().getWindow()).setMinWidth(375);
-        keyBindingsDialog.show();
+public class KeyBindingsDialogView extends Dialog<Void> {
+
+    @FXML private TreeTableView<KeyBindingViewModel> keyBindingsTable;
+    @FXML private TreeTableColumn<KeyBindingViewModel, String> actionColumn;
+    @FXML private TreeTableColumn<KeyBindingViewModel, String> shortcutColumn;
+    @FXML private TreeTableColumn<KeyBindingViewModel, String> resetColumn;
+
+    @Inject private KeyBindingRepository keyBindingRepository;
+    @Inject private DialogService dialogService;
+    @Inject private PreferencesService preferences;
+    private KeyBindingsDialogViewModel viewModel;
+
+    public KeyBindingsDialogView() {
+        this.setTitle(Localization.lang("Key bindings"));
+        this.setResizable(true);
+
+        Stage stage = (Stage) this.getDialogPane().getScene().getWindow();
+        stage.setMinHeight(475);
+        stage.setMinWidth(375);
+
+        Parent content = ViewLoader.view(this)
+                                   .load()
+                                   .getView();
+        this.getDialogPane().setContent(content);
     }
 
+    @FXML
+    private void initialize() {
+        viewModel = new KeyBindingsDialogViewModel(keyBindingRepository, dialogService, preferences);
+
+        keyBindingsTable.getSelectionModel().setSelectionMode(SelectionMode.SINGLE);
+        viewModel.selectedKeyBindingProperty().bind(
+                EasyBind.monadic(keyBindingsTable.selectionModelProperty())
+                        .flatMap(SelectionModel::selectedItemProperty)
+                        .selectProperty(TreeItem::valueProperty)
+        );
+        keyBindingsTable.setOnKeyPressed(evt -> viewModel.setNewBindingForCurrent(evt));
+        keyBindingsTable.rootProperty().bind(
+                EasyBind.map(viewModel.rootKeyBindingProperty(),
+                        keybinding -> new RecursiveTreeItem<>(keybinding, KeyBindingViewModel::getChildren))
+        );
+        actionColumn.setCellValueFactory(cellData -> cellData.getValue().getValue().nameProperty());
+        shortcutColumn.setCellValueFactory(cellData -> cellData.getValue().getValue().shownBindingProperty());
+        resetColumn.setCellFactory(new ViewModelTreeTableCellFactory<KeyBindingViewModel, String>()
+                .withGraphic(keyBinding -> keyBinding.getIcon().map(JabRefIcon::getGraphicNode).orElse(null))
+                .withOnMouseClickedEvent(keyBinding -> evt -> keyBinding.resetToDefault())
+        );
+    }
+
+    @FXML
+    private void closeDialog() {
+        close();
+    }
+
+    @FXML
+    private void saveKeyBindingsAndCloseDialog() {
+        viewModel.saveKeyBindings();
+        closeDialog();
+    }
+
+    @FXML
+    private void setDefaultBindings() {
+        viewModel.resetToDefault();
+    }
 }
