+++ src/main/resources/META-INF/services/com.airhacks.afterburner.views.ResourceLocator
@@ -1,0 +1,1 @@
+org.jabref.logic.l10n.LocalizationLocator
